convert sheet to navlink view and move structs to new files
